Added Serp Popup Enhancement (#1060)
* fix: fixed the bug with popup in small screens

* feat: added table of similar queries
