+++ admin/src/assets/styles/components/_SerpPanel.scss
@@ -16,17 +16,8 @@
     }
 
     &-keywords-item {
-        display: inline-block;
-        padding: 0.5rem;
-        border: 1px solid $grey-light;
-        border-radius: 10px;
-        margin-right: 0.5rem;
         cursor: pointer;
-
-        &:hover {
-            text-decoration: underline;
-        }
-
+        text-decoration: underline;
     }
 
     &-input {
