+++ admin/src/lib/serpQueries.js
@@ -7,8 +7,8 @@ export const getTopUrls = async ( query, domain_type, limit ) => {
 	}
 };
 
-export const getQueryClusterKeywords = async ( query, max_position = 10, competitors = 4 ) => {
-	const res = await postFetch( 'serp-queries/query-cluster', { query, max_position, competitors } );
+export const getQueryClusterKeywords = async ( query, max_position = 10, competitors = 4, with_stats = false ) => {
+	const res = await postFetch( 'serp-queries/query-cluster', { query, max_position, competitors, with_stats } );
 	if ( res.ok ) {
 		return await res.json();
 	}
